+++ core/src/test/java/ai/greycos/solver/core/impl/score/stream/common/inliner/HardSoftScoreInlinerTest.java
@@ -84,14 +84,13 @@ class HardSoftScoreInlinerTest
   void impactAllMatchWeightOverflow() {
     var constraintWeight = HardSoftScore.of(10, 100);
     var impacter = buildScoreImpacter(constraintWeight);
-    assertThatThrownBy(
-            () -> impacter.impactScore(Integer.MAX_VALUE, ConstraintMatchSupplier.empty()))
+    assertThatThrownBy(() -> impacter.impactScore(Long.MAX_VALUE, ConstraintMatchSupplier.empty()))
         .isInstanceOf(ArithmeticException.class);
   }
 
   @Test
   void impactAllTotalOverflow() {
-    var constraintWeight = HardSoftScore.of(Integer.MAX_VALUE, Integer.MAX_VALUE);
+    var constraintWeight = HardSoftScore.of(Long.MAX_VALUE, Long.MAX_VALUE);
     var impacter = buildScoreImpacter(constraintWeight);
     impacter.impactScore(1, ConstraintMatchSupplier.empty());
     assertThatThrownBy(() -> impacter.impactScore(1, ConstraintMatchSupplier.empty()))
